first working version of kk-ui.sh

+++ kk-ui.sh
@@ -19,26 +19,13 @@ clean_env() { # Remove all the intermediate files
 	local i
 
 	for i in `echo $APPLIST`; do # Remove old decompiled directories
-		if [ -d $i ] ; then
-			rm -rf $i
-		fi
+		[ -e $i ] && rm -rf $i
 	done
 
-	if find dist -name *.apk 1>/dev/null ; then # Remove old built apks
-		rm `find dist -name *.apk`
-	fi
-
-	if ls *.apk 1>/dev/null ; then # Remove straggaling apks
-		rm *.apk
-	fi
-
-	if [ -e src/$BUILD ] ; then # Remove extracted apks
-		rm -rf src/$BUILD
-	fi
-
-	if ls -d src/pa-4.* 1>/dev/null ; then # Find a better way to do this
-		rm -rf src/pa-4.*
-	fi
+	[ "`find dist -name \*.apk`" ] && rm `find dist -name \*.apk` # Cleanout the dist folder
+	[ "`find . -maxdepth 1 -name \*.apk`" ] && rm *.apk # remove any straggling .apk files (should never be any)
+	[ -e src/$BUILD ] && rm -rf src/$BUILD # remove extracted zips from this build
+	[ "`find src -maxdepth 1 -name pa_d2lte\*`" ] && rm -rf src/pa_d2lte* # remove extracted zips from old builds (find a better way)
 }
 
 check_deps() { # Make sure we have everything we need
@@ -55,19 +42,19 @@ setup_env() { # Get things ready to build
 	# Extract original .apk files
 	mkdir src/$BUILD
 	for i in `echo $PRIVAPP`; do
-		unzip -jd src/$BUILD $ROM system/priv-app/$i.apk
+		unzip -jd src/$BUILD $ROM system/priv-app/$i.apk 1>&2
 	done
-	unzip -jd src/$BUILD $ROM system/framework/framework-res.apk
+	unzip -jd src/$BUILD $ROM system/framework/framework-res.apk 1>&2
 	for i in `echo $SYSAPP`; do
-		unzip -jd src/$BUILD $ROM system/app/$i.apk
+		unzip -jd src/$BUILD $ROM system/app/$i.apk 1>&2
 	done
 
 	# Install the framework. apktool 1.5.3 and 2.0.0-rc1 use different syntax for this
 	# TODO do this better. See Comment at var definitions
-	if [ "$APKTOOL" == "$APKTOOL_STABLE" ] ; then
-		$APKTOOL -t $TAG if src/$BUILD/framework-res.apk
-	elif [ "$APKTOOL" == "$APKTOOL_TRDS" ] ; then
-		$APKTOOL if src/$BUILD/framework-res.apk $TAG
+	if [ "$APKTOOL" = "$APKTOOL_STABLE" ] ; then
+		$APKTOOL -t $TAG if src/$BUILD/framework-res.apk 1>&2
+	elif [ "$APKTOOL" = "$APKTOOL_TRDS" ] ; then
+		$APKTOOL if src/$BUILD/framework-res.apk $TAG 1>&2
 	else
 		echo 'ERROR: Unknown apktool. Cannot install framework.' 1>&2
 		return $E_GEN
@@ -79,32 +66,45 @@ decode() { # $1=package.akp (include path!)
 		echo 'ERROR: No package given to decode.' 1>&2
 		return $E_GEN
 	fi
-	$APKTOOL d -t $TAG $1
+	$APKTOOL d -t $TAG $1 1>&2
 }
 
 build() { # $1=package (string); $2=special build conditions (dialer,gallery)
 	local SIG='META-INF/MANIFEST.MF META-INF/CERT.SF META-INF/CERT.RSA AndroidManifest.xml'
-	local FILE
 	local FILTER="has no default translation"
+	local FILE
+	local i
 
 	if [ -z $1 ] ; then
 		echo 'ERROR: No packagae given to build.' 1>&2
 		return $E_GEN
 	fi
 
-	if [ $2 == "dialer"]; then #TODO find out what acutally is breaking this rather than doing this bit of overkill.
+	if [[ "$1" == "TeleService" || "$1" == *Dialer* ]] ; then #TODO find out what acutally is breaking this rather than doing this bit of overkill.
 		for i in `ls $1/res/values-*/strings.xml -d`; do
 			sed -i '/name="throttle_time_frame_subtext"/ s/text">/text" formatted="false">/' $i
 		done
-	elif [ $2 == "gallery" ]; then 
+	elif [[ "$1" == *Gallery* || "$1" == *Camera* ]] ; then 
 		FILE=$1/res/values/styles.xml
 		sed -i '/name="TextAppearance.DialogWindowTitle"/ s/Title">/Title" parent="">/' $FILE
 	fi
 
-	echo 'Building '$i'.apk' 1>&2
 	$APKTOOL -t $TAG b $1 2>&1 | grep -v $FILTER 1>&2
-	unzip -od $1/build/apk src/$BUILD/$i.apk $SIG
+	unzip -od $1/build/apk src/$BUILD/$1.apk `echo $SIG`
 	$APKTOOL -t $TAG b $1 2>&1 | grep -v $FILTER 1>&2
+}
+
+recode() { # build and decode in place
+	build $1
+	E_BUILD_FAIL=$?
+	if [ $E_BUILD_FAIL -eq 0 ] ; then
+		mv $1/dist/$1.apk /tmp
+		rm -r $1
+		decode /tmp/$1.apk
+		rm /tmp/$1.apk
+	else
+		return $E_BUILD_FAIL
+	fi
 }
 
 copy_res() {
@@ -114,11 +114,15 @@ copy_res() {
 		return $E_GEN
 	fi
 
+	if [ ! -d $RESDIR/$1/res ] ; then
+		echo "WARNING: No new resources for $1"
+		return $W_NORES
+	fi
+
 	for i in `ls $RESDIR/$1/res`; do
-		if [ -d $1/res/$i ]; then
-			cp -r $RESDIR/$1/res/$i $1/res
-		fi
+		[ -d $1/res/$i ] && cp -r $RESDIR/$1/res/$i $1/res
 	done
+
 }
 
 remove_holo_blue() {
@@ -129,7 +133,7 @@ remove_holo_blue() {
 
 	decode src/$BUILD/$1.apk
 	copy_res $1
-	sedit_$1
+	type sedit_$1 1>/dev/null && sedit_$1 # Not all apks need edited
 	build $1
 }
 
@@ -146,22 +150,25 @@ package() { #TODO cleanup
 	done
 	cp src/Follow_KK_UI.zip dist/${BUILD}_follow-kk-ui.zip
 	cd dist
-	zip -ur ${BUILD}_follow-kk-ui.zip $OUTDIR
-	rm `find system -name *.apk`
+	zip -ur ${BUILD}_follow-kk-ui.zip system
+	mv ${BUILD}_follow-kk-ui.zip $OUTDIR
+	rm `find system -name \*.apk`
+	cd $BASEDIR
 
 	#build reversion .zip
 	for i in `echo $PRIVAPP`; do
-		 cp ../src/$BUILD/$i.apk system/priv-app
+		 cp src/$BUILD/$i.apk dist/system/priv-app
 	done
-	mv ../src/$BUILD/framework-res.apk system/framework
+	mv src/$BUILD/framework-res.apk dist/system/framework
 	for i in `echo $SYSAPP`; do
-		cp ../src/$BUILD/$i.apk system/app
+		cp src/$BUILD/$i.apk dist/system/app
 	done
-	cp ../src/Follow_KK_UI.zip ${BUILD}_revert.zip
+	cp src/Follow_KK_UI.zip dist/${BUILD}_revert.zip
+	cd dist
 	zip -ur ${BUILD}_revert.zip system
 	mv ${BUILD}_revert.zip $OUTDIR
-
-	cd ..
+	rm `find system -name \*.apk`
+	cd $BASEDIR
 }
 
 sedit_Calculator() {
@@ -212,14 +219,14 @@ sedit_framework-res() {
 	sed -i '/TextView/ s/textColor="#80000000"/textColor="#80ffffff"/' $FILE
 
 	FILE=framework-res/res/values/colors.xml
-	sed -i '$ i\
+	grep default_dialog_text_holo_dark $FILE 1>/dev/null || sed -i '$ i\
     <color name="default_dialog_text_holo_dark">@android:color/background_holo_light</color>\
     <color name="default_dialog_text_holo_light">@android:color/background_holo_dark</color>\
     <color name="default_dialog_divider_holo_dark">@android:color/background_holo_light</color>\
-    <color name="default_dialog_divider_holo_light">#ffacacac</color>' $FILE
-	sed -i '/holo_blue_light/ s/33b5e5/dbdbdb/' $FILE
-	sed -i '/highlighted_text_holo_dark/ s/33b5e5/ffffff/' $FILE
-	sed -i '/highlighted_text_holo_light/ s/33b5e5/000000/' $FILE
+    <color name="default_dialog_divider_holo_light">#ffacacac</color>' $FILE # It may end up being necessary to check for each color, and also to make an edit enforcing the right value, if it is already defined
+	sed -i '/name="holo_blue_light"/ s/33b5e5/dbdbdb/' $FILE
+	sed -i '/name="highlighted_text_holo_dark"/ s/33b5e5/ffffff/' $FILE
+	sed -i '/name="highlighted_text_holo_light"/ s/33b5e5/000000/' $FILE
 
 	FILE=framework-res/res/values/styles.xml
 	sed -i '/name="TextAppearance.Holo.DialogWindowTitle"/,/textColor/ s/holo_blue_light/default_dialog_text_holo_dark/' $FILE
@@ -229,9 +236,11 @@ sedit_framework-res() {
         <item name="maxHeight">16.0dip</item>\
         <item name="indeterminateDrawable">@drawable/progress_indeterminate_horizontal_holo_light</item>' $FILE
 	sed -i '/name="Widget.Holo.Light.ProgressBar.Horizontal"/,/<\/style>/ s/<\/style>/    <item name="minHeight">16.0dip<\/item>\
-    <\/style>/' $FILE
+    <\/style>/' $FILE # Surely, there is a more elegant way of doing this.
 	sed -i '/name="Widget.Holo.Light.SeekBar"/,/<\/style>/ s/<\/style>/    <item name="android:thumb">@android:drawable\/scrubber_control_selector_holo_light<\/item>\
-    <\/style>/' $FILE
+    <\/style>/' $FILE # And this.
+
+	# http://forum.xda-developers.com/showpost.php?p=51444245&postcount=475
 	sed -i '/name="Widget.Holo.Light.TabWidget"/ s/ \/>/>/' $FILE
 	sed -i '/name="Widget.Holo.Light.TabWidget"/ a\
         <item name="divider">?dividerVertical</item>\
@@ -241,8 +250,18 @@ sedit_framework-res() {
         <item name="measureWithLargestChild">true</item>\
         <item name="showDividers">middle</item>\
         <item name="dividerPadding">8.0dip</item>\
-        <item name="tabLayout">@layout/tab_indicator_holo</item>\
+        <item name="tabLayout">@layout/tab_indicator_holo_light</item>\
     </style>' $FILE
+	sed -i '/name="Widget.Holo.Light.Tab"/,/<\/style>/ s/tab_indicator_holo/tab_indicator_holo_light/' $FILE
+	sed -i '/name="Widget.Holo.Light.ActionBar.TabView"/ s/ \/>/>/' $FILE
+	sed -i '/name="Widget.Holo.Light.ActionBar.TabView"/ a\
+        <item name="background">@drawable/tab_indicator_ab_holo_light</item>\
+        <item name="paddingStart">16.0dip</item>\
+        <item name="paddingEnd">16.0dip</item>\
+    </style>' $FILE
+
+	# http://forum.xda-developers.com/showpost.php?p=52502121&postcount=593
+	sed -i '/name="Widget.Holo.Light.TabWidget"/,/<\/style>/ s/tab_indicator_holo_light/tab_indicator_holo/' $FILE
 	LINE=`sed -n '/name="Widget.Holo.Light.Tab"/=' $FILE`
 	sed -i '/name="Widget.Holo.Light.Tab"/,/<\/style>/ d' $FILE
 	sed -i $LINE' i\    <style name="Widget.Holo.Light.Tab" parent="@style\/Widget.Holo.Tab" \/>' $FILE
@@ -253,18 +272,13 @@ sedit_framework-res() {
         <item name="tabIndicatorHolo">@drawable/tab_indicator_holo_light</item>' $FILE
 	sed -i '/name="Theme.Holo.Light.DarkActionBar"/ a\
         <item name="tabIndicatorHolo">@drawable/tab_indicator_holo_light</item>' $FILE
-	sed -i '/name="Widget.Holo.Light.ActionBar.TabView"/ s/Holo.ActionBar.TabView" \/>/ActionBar.TabView">/' $FILE
-	sed -i '/name="Widget.Holo.Light.ActionBar.TabView"/ a\
-        <item name="background">@drawable/tab_indicator_ab_holo_light</item>\
-        <item name="paddingStart">16.0dip</item>\
-        <item name="paddingEnd">16.0dip</item>\
-    </style>' $FILE
+
 	FILE=framework-res/res/values/attrs.xml
 	sed -i '/<resources>/ a\
     <attr name="tabIndicatorHolo" format="reference" />' $FILE
 
 	FILE=framework-res/res/layout/tab_indicator_holo_light.xml
-	rm $FILE
+	[ -e $FILE ] && rm $FILE
 
 	FILE=framework-res/res/values/drawables.xml
 	sed -i '/name="notification_template_icon_bg"/ s/3333b5e5/00000000/' $FILE
@@ -275,25 +289,25 @@ sedit_framework-res() {
 
 sedit_SystemUI() {
 	local FILE=SystemUI/res/layout/status_bar_no_recent_apps.xml
-	sed -i 's/holo_blue_light/white/' $FILE
+	sed -i 's/holo_blue_light/white/g' $FILE
 }
 
 sedit_Dialer() {
 	local FILE=Dialer/res/values/colors.xml 
-	sed -i '/"incall_call_banner_background"/ s/000000</e5e5e5</' $FILE
-	sed -i '/"dialpad_background"/ s/000000/e5e5e5/' $FILE
-	sed -i '/"incall_secondary_info_background"/ s/33b5e5/e5e5e5/' $FILE
-	sed -i '/"incall_call_banner_text_color"/ s/ffffff</404040</' $FILE
-	sed -i '/"dtmf_dialer_display_text"/ s/ffffff</404040</' $FILE
-	sed -i '/"dialtacts_primary_text_color"/ s/000000/111111/' $FILE
-	sed -i '/"dialpad_text_color"/ s/ffffff</404040</' $FILE
-	sed -i '/"call_log_voicemail_highlight_color"/ s/33b5e5/404040/' $FILE
-	sed -i '/"background_dialer_list_items"/ s/eeeeee/f5f5f5/' $FILE
-	sed -i '/"background_dialpad"/ s/#.*</#ffe5e5e5</' $FILE
-	sed -i '/"dialpad_primary_text_color"/ s/#.*</#ff404040</' $FILE
-	sed -i '/"dialpad_secondary_text_color"/ s/#.*</#ffaaaaaa</' $FILE
-	sed -i '/"actionbar_background_color"/ s/ffffff</e6e6e6</' $FILE
-	sed -i '/"secondary_text_color"/ s/888888/777777/' $FILE
+	sed -i '/name="incall_call_banner_background"/ s/000000</e5e5e5</' $FILE
+	sed -i '/name="dialpad_background"/ s/000000/e5e5e5/' $FILE
+	sed -i '/name="incall_secondary_info_background"/ s/33b5e5/e5e5e5/' $FILE
+	sed -i '/name="incall_call_banner_text_color"/ s/ffffff</404040</' $FILE
+	sed -i '/name="dtmf_dialer_display_text"/ s/ffffff</404040</' $FILE
+	sed -i '/name="dialtacts_primary_text_color"/ s/000000/111111/' $FILE
+	sed -i '/name="dialpad_text_color"/ s/ffffff</404040</' $FILE
+	sed -i '/name="call_log_voicemail_highlight_color"/ s/33b5e5/404040/' $FILE
+	sed -i '/name="background_dialer_list_items"/ s/eeeeee/f5f5f5/' $FILE
+	sed -i '/name="background_dialpad"/ s/#.*</#ffe5e5e5</' $FILE
+	sed -i '/name="dialpad_primary_text_color"/ s/#.*</#ff404040</' $FILE
+	sed -i '/name="dialpad_secondary_text_color"/ s/#.*</#ffaaaaaa</' $FILE
+	sed -i '/name="actionbar_background_color"/ s/ffffff</e6e6e6</' $FILE
+	sed -i '/name="secondary_text_color"/ s/888888/777777/' $FILE
 
 	FILE=Dialer/res/values/drawables.xml
 	sed -i '/name="grayBg"/ s/333333/e5e5e5/' $FILE
@@ -311,13 +325,10 @@ sedit_Camera2() {
 	sed -i '/name="holo_blue_light"/ s/33b5e5/888888/' $FILE
 	sed -i '/name="pano_progress_done"/ s/33525e/888888/' $FILE
 	sed -i '/name="pano_progress_indication"/ s/0099cc/ffffff/' $FILE
-	sed -i '$ i\
+	grep 'name="pie_selected_paint_color"' $FILE || sed -i '$ i\
     <color name="pie_selected_paint_color">#88888888</color>' $FILE
 	
-	build Camera2
-	mv Camera2/dist/Camera2.apk .
-	decode ./Camera2.apk
-	rm ./Camera2.apk
+	recode ./Camera2
 	PIECOLOR=`grep pie_selected_paint_color Camera2/res/values/public.xml | cut -f6 -d'"'`
 	FILE=Camera2/smali/com/android/camera/ui/PieRenderer.smali
 	sed -i '/.line 198/,+14 d' $FILE
@@ -344,6 +355,7 @@ sedit_Mms() {
 }
 
 # Definitions
+OLDDIR=`pwd`
 BASEDIR=~/apktool
 OUTDIR=~/Dropbox/android/d2vzw/kk-theme
 RESDIR=src/resources
@@ -362,12 +374,12 @@ APKTOOL=$APKTOOL_STABLE
 
 REVERT=false
 VERBOSE=false
+W_NORES=10
 E_NORES=10 # Res files not found
 E_GEN=74 # General error (probably a fault in script)
 unset ROM
-unset ROMDIR
 unset BUILD
-
+unset LOGFILE
 
 # Get options
 while :
@@ -375,6 +387,7 @@ do
 	case $1 in
 		-d)
 			OUTDIR=$2
+			#TODO check and adjust for rel/full paths
 			shift
 			;;
 		-h)
@@ -397,13 +410,21 @@ do
 		*.zip)
 			ROM=$1
 			BUILD=`basename ${ROM%.*}`
+			TAG=`echo $BUILD | cut -d\- -f4`
 			;;
 		*) # No more options. Stop while loop.
+			if [[ -z "$ROM" ]] ; then
+				echo "No ROM file given!"
+				show_help
+				exit 1
+			fi
 			break
 			;;
 	esac
 	shift
 done
+
+cd $BASEDIR
 
 check_deps
 clean_env
@@ -414,3 +435,4 @@ for APP in `echo $APPLIST`; do
 done
 package
 echo 'Dropbox link:\nhttps://www.dropbox.com/sh/3azn7pw2vkcrhan/URyQaVww7v'
+cd $OLDDIR
